requirements.txt
aiogram==3.20.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1